fix(game): cast the window height for the photo-mode FOV refresh
MSVC /W4 /WX rejects the implicit int -> float conversion (C4244); GCC's
-Wall -Wextra -Wpedantic does not, so the Linux build was clean and the Windows
leg failed. The live-session call site three hundred lines up already casts --
this one was written without it.

Same shape as the C4456 shadow that failed Stage 1's Windows leg: a warning
class MSVC treats as an error and the Linux flags do not mention at all. The
pre-push sweep now covers -Wfloat-conversion alongside -Wshadow.

+++ game/fighters-legacy/Game.cpp
@@ -2566,7 +2566,8 @@ void Game::run() {
                 // Keep terrain refinement honest about the FOV in use, or SSE tile selection refines for
                 // a field of view the player is not looking through.
                 if (d.services.terrainStreamer)
-                    d.services.terrainStreamer->setViewParams(d.services.p.window->height(), glm::radians(fovDeg));
+                    d.services.terrainStreamer->setViewParams(static_cast<float>(d.services.p.window->height()),
+                                                              glm::radians(fovDeg));
 
                 if (d.services.rendererSettings.evOffset != d.services.photo.evOffset) {
                     d.services.rendererSettings.evOffset = d.services.photo.active ? d.services.photo.evOffset : 0.f;
